a 21 de septiembre 22

+++ src/teoriaInterfacesFuncionales00/PruebaACtionFunctionalParametrized07.java
@@ -3,7 +3,8 @@ package teoriaInterfacesFuncionales00;
 public class PruebaACtionFunctionalParametrized07 {
 public static void main(String[] args) {
 	ACtionFunctionalPArametrized06<Integer,String> action=(b)->{return Integer.valueOf(b);};
-	int valor=action.convert("7");
+	float valor=action.convert("7");
 	System.out.println(valor);
+	
  }
 }
